Check For 401 Status Rather Than !req.ok
Previously, the login fetch request checks for !req.ok. Currently, this
is true on successful login as there is a 404 response status. While
this is the case, the check was changed to check for 401 response
status. This avoids triggering the if statement, meaning that state
errors are not set and that the user is redirected to the home page.